requirements.txt
pipecat-ai==0.0.54
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
redis==5.2.1
qdrant-client==1.12.1
langfuse==2.58.1
python-dotenv==1.0.1
